+++ order-api/src/main/java/com/zerobase/cms/order/application/CartApplication.java
@@ -5,8 +5,8 @@ import com.zerobase.cms.order.domain.model.ProductItem;
 import com.zerobase.cms.order.domain.product.AddProductCartForm;
 import com.zerobase.cms.order.domain.redis.Cart;
 import com.zerobase.cms.order.exception.CustomException;
-import com.zerobase.cms.order.service.CartService;
-import com.zerobase.cms.order.service.ProductSearchService;
+import com.zerobase.cms.order.service.CartServiceImpl;
+import com.zerobase.cms.order.service.ProductSearchServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -22,8 +22,8 @@ import static com.zerobase.cms.order.exception.ErrorCode.NO_EXIST_PRODUCT;
 @Service
 @RequiredArgsConstructor
 public class CartApplication {
-    private final CartService cartService;
-    private final ProductSearchService productSearchService;
+    private final CartServiceImpl cartService;
+    private final ProductSearchServiceImpl productSearchService;
 
     public Cart addCart(Long customerId, AddProductCartForm form) {
         Product product = productSearchService.getByProductId(form.getId());
