+++ order-api/src/main/java/com/zerobase/cms/order/service/ProductItemService.java
@@ -4,61 +4,18 @@ import com.zerobase.cms.order.domain.model.Product;
 import com.zerobase.cms.order.domain.model.ProductItem;
 import com.zerobase.cms.order.domain.product.AddProductItemForm;
 import com.zerobase.cms.order.domain.product.UpdateProductItemForm;
-import com.zerobase.cms.order.domain.repository.ProductItemRepository;
-import com.zerobase.cms.order.domain.repository.ProductRepository;
-import com.zerobase.cms.order.exception.CustomException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.zerobase.cms.order.exception.ErrorCode.*;
-
-@Service
-@RequiredArgsConstructor
-public class ProductItemService {
-    private final ProductRepository productRepository;
-    private final ProductItemRepository productItemRepository;
+public interface ProductItemService {
+    @Transactional
+    Product addProductItem(Long sellerId, AddProductItemForm form);
 
     @Transactional
-    public Product addProductItem(Long sellerId, AddProductItemForm form) {
-        Product product = productRepository.findBySellerIdAndId(sellerId, form.getProductId())
-                .orElseThrow(() -> new CustomException(NO_EXIST_PRODUCT));
-
-        validateAddProductItem(product, form);
-
-        ProductItem productItem = ProductItem.of(sellerId, form);
-        product.getProductItems().add(productItem);
-        return product;
-    }
+    ProductItem updateProductItem(Long sellerId, UpdateProductItemForm form);
 
     @Transactional
-    public ProductItem updateProductItem(Long sellerId, UpdateProductItemForm form) {
-        ProductItem productItem = productItemRepository.findById(form.getId())
-                .filter(pi -> pi.getSellerId().equals(sellerId))
-                .orElseThrow(() -> new CustomException(NO_EXIST_ITEM));
-        productItem.setName(form.getName());
-        productItem.setCount(form.getCount());
-        productItem.setPrice(form.getPrice());
-        return productItem;
-    }
+    void deleteProductItem(Long sellerId, Long productItemId);
 
     @Transactional
-    public void deleteProductItem(Long sellerId, Long productItemId) {
-        ProductItem productItem = productItemRepository.findById(productItemId)
-                .filter(pi -> pi.getSellerId().equals(sellerId))
-                .orElseThrow(() -> new CustomException(NO_EXIST_ITEM));
-        productItemRepository.delete(productItem);
-    }
-
-    @Transactional
-    public ProductItem getProductItem(Long id) {
-        return productItemRepository.getReferenceById(id);
-    }
-
-    private void validateAddProductItem(Product product, AddProductItemForm form) {
-        if (product.getProductItems().stream()
-                .anyMatch(item -> item.getName().equals(form.getName()))) {
-            throw new CustomException(SAME_ITEM_NAME);
-        }
-    }
+    ProductItem getProductItem(Long id);
 }
