+++ order-api/src/main/java/com/zerobase/cms/order/service/ProductSearchService.java
@@ -1,30 +1,13 @@
 package com.zerobase.cms.order.service;
 
 import com.zerobase.cms.order.domain.model.Product;
-import com.zerobase.cms.order.domain.repository.ProductRepository;
-import com.zerobase.cms.order.exception.CustomException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-import static com.zerobase.cms.order.exception.ErrorCode.NO_EXIST_PRODUCT;
+public interface ProductSearchService {
+    List<Product> searchByName(String name);
 
-@Service
-@RequiredArgsConstructor
-public class ProductSearchService {
-    private final ProductRepository productRepository;
+    Product getByProductId(Long productId);
 
-    public List<Product> searchByName(String name) {
-        return productRepository.searchByName(name);
-    }
-
-    public Product getByProductId(Long productId) {
-        return productRepository.findWithProductItemsById(productId)
-                .orElseThrow(() -> new CustomException(NO_EXIST_PRODUCT));
-    }
-
-    public List<Product> getListByProductIds(List<Long> productIds) {
-        return productRepository.findAllByIdIn(productIds);
-    }
+    List<Product> getListByProductIds(List<Long> productIds);
 }
