+++ order-api/src/main/java/com/zerobase/cms/order/service/ProductSearchServiceImpl.java
@@ -1,0 +1,33 @@
+package com.zerobase.cms.order.service;
+
+import com.zerobase.cms.order.domain.model.Product;
+import com.zerobase.cms.order.domain.repository.ProductRepository;
+import com.zerobase.cms.order.exception.CustomException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+import static com.zerobase.cms.order.exception.ErrorCode.NO_EXIST_PRODUCT;
+
+@Service
+@RequiredArgsConstructor
+public class ProductSearchServiceImpl implements ProductSearchService {
+    private final ProductRepository productRepository;
+
+    @Override
+    public List<Product> searchByName(String name) {
+        return productRepository.searchByName(name);
+    }
+
+    @Override
+    public Product getByProductId(Long productId) {
+        return productRepository.findWithProductItemsById(productId)
+                .orElseThrow(() -> new CustomException(NO_EXIST_PRODUCT));
+    }
+
+    @Override
+    public List<Product> getListByProductIds(List<Long> productIds) {
+        return productRepository.findAllByIdIn(productIds);
+    }
+}
