+++ user-api/src/main/java/com/zerobase/cms/user/controller/CustomerController.java
@@ -5,8 +5,8 @@ import com.zerobase.cms.user.domain.customer.CustomerDto;
 import com.zerobase.cms.user.domain.model.Customer;
 import com.zerobase.cms.user.exception.CustomException;
 import com.zerobase.cms.user.exception.ErrorCode;
-import com.zerobase.cms.user.service.customer.CustomerBalanceService;
-import com.zerobase.cms.user.service.customer.CustomerService;
+import com.zerobase.cms.user.service.customer.CustomerBalanceServiceImpl;
+import com.zerobase.cms.user.service.customer.CustomerServiceImpl;
 import com.zerobase.domain.common.UserVo;
 import com.zerobase.domain.config.JwtAuthenticationProvider;
 import lombok.RequiredArgsConstructor;
@@ -18,8 +18,8 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/customer")
 public class CustomerController {
     private final JwtAuthenticationProvider provider;
-    private final CustomerService customerService;
-    private final CustomerBalanceService customerBalanceService;
+    private final CustomerServiceImpl customerService;
+    private final CustomerBalanceServiceImpl customerBalanceService;
 
     @GetMapping("/getInfo")
     public ResponseEntity<CustomerDto> getInfo(@RequestHeader(name = "X-AUTH-TOKEN") String token) {
