+++ user-api/src/main/java/com/zerobase/cms/user/service/customer/CustomerServiceImpl.java
@@ -1,0 +1,32 @@
+package com.zerobase.cms.user.service.customer;
+
+import com.zerobase.cms.user.domain.model.Customer;
+import com.zerobase.cms.user.domain.repository.CustomerRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class CustomerServiceImpl implements CustomerService {
+    private final CustomerRepository customerRepository;
+
+    @Override
+    public Optional<Customer> findByIdAndEmail(Long id, String email) {
+        return customerRepository.findById(id)
+                .stream()
+                .filter(customer -> customer.getEmail().equals(email))
+                .findFirst();
+    }
+
+    @Override
+    public Optional<Customer> findValidCustomer(String email, String password) {
+        return customerRepository.findByEmail(email)
+                .stream()
+                .filter(customer ->
+                        customer.getPassword().equals(password) && customer.isVerify())
+                .findFirst();
+
+    }
+}
