+++ user-api/src/main/java/com/zerobase/cms/user/service/seller/SellerService.java
@@ -2,74 +2,23 @@ package com.zerobase.cms.user.service.seller;
 
 import com.zerobase.cms.user.domain.SignUpForm;
 import com.zerobase.cms.user.domain.model.Seller;
-import com.zerobase.cms.user.domain.repository.SellerRepository;
-import com.zerobase.cms.user.exception.CustomException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-import static com.zerobase.cms.user.exception.ErrorCode.*;
-import static java.util.Locale.ROOT;
+public interface SellerService {
+    Optional<Seller> findByIdAndEmail(Long id, String email);
 
-@Service
-@RequiredArgsConstructor
-public class SellerService {
-    private final SellerRepository sellerRepository;
+    Optional<Seller> findValidSeller(String email, String password);
 
-    public Optional<Seller> findByIdAndEmail(Long id, String email) {
-        return sellerRepository.findByIdAndEmail(id, email);
-    }
+    Seller signUp(SignUpForm form);
 
-    public Optional<Seller> findValidSeller(String email, String password) {
-        return sellerRepository.findByEmailAndPasswordAndVerifyIsTrue(email, password);
-    }
-
-    public Seller signUp(SignUpForm form) {
-        return sellerRepository.save(Seller.from(form));
-    }
-
-    public boolean isEmailExist(String email) {
-        return sellerRepository.findByEmail(email.toLowerCase(ROOT))
-                .isPresent();
-    }
+    boolean isEmailExist(String email);
 
     @Transactional
-    public void verifyEmail(String email, String code) {
-        Seller seller = sellerRepository.findByEmail(email.toLowerCase(ROOT))
-                .orElseThrow(() -> new CustomException(NO_EXIST_USER));
-
-        validateVerifyEmail(code, seller);
-
-        seller.setVerify(true);
-    }
+    void verifyEmail(String email, String code);
 
     @Transactional
-    public LocalDateTime changeSellerValidateEmail(Long sellerId, String verificationCode) {
-        Optional<Seller> sellerOptional = sellerRepository.findById(sellerId);
-        if (sellerOptional.isPresent()) {
-            Seller seller = sellerOptional.get();
-            seller.setVerificationCode(verificationCode);
-            seller.setVerifyExpiredAt(LocalDateTime.now().plusDays(1));
-            return seller.getVerifyExpiredAt();
-        } else {
-            throw new CustomException(NO_EXIST_USER);
-        }
-    }
-
-    private void validateVerifyEmail(String code, Seller seller) {
-        if (seller.isVerify()) {
-            throw new CustomException(ALREADY_VERIFY);
-        }
-
-        if (!seller.getVerificationCode().equals(code)) {
-            throw new CustomException(WRONG_VERIFICATION);
-        }
-
-        if (seller.getVerifyExpiredAt().isBefore(LocalDateTime.now())) {
-            throw new CustomException(EXPIRE_CODE);
-        }
-    }
+    LocalDateTime changeSellerValidateEmail(Long sellerId, String verificationCode);
 }
